requirements.txt
discord.py>=2.0.0
discord-ext-menus
aiohttp
requests
psutil
Pillow
tasksio
jishaku
wavelink
discord-ext-menus
setuptools
httpx
colorama
